% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bordeauxNA}
\alias{bordeauxNA}
\title{Quality of wine dataset}
\format{
A data frame with 34 observations on the following 5 variables.
\describe{ \item{Temperature}{a numeric vector}
\item{Sunshine}{a numeric vector} \item{Heat}{a numeric
vector} \item{Rain}{a numeric vector} \item{Quality}{an
ordered factor with levels \code{1} < \code{2} < \code{3}} }
}
\source{
P. Bastien, V. Esposito-Vinzi, and M. Tenenhaus. (2005). PLS
generalised linear regression. \emph{Computational Statistics & Data
Analysis}, 48(1):17-46.
}
\description{
Quality of Bordeaux wines (\code{Quality}) and four potentially predictive
variables (\code{Temperature}, \code{Sunshine}, \code{Heat} and
\code{Rain}).
}
\details{
The value of x1 for the first observation was removed from the matrix of
predictors on purpose.

The \code{bordeauxNA} is a dataset with a missing value for testing purpose.
}
\examples{

data(bordeauxNA)
str(bordeauxNA)

}
\references{
M. Tenenhaus. (2005). La regression logistique PLS. In J.-J.
Droesbeke, M. Lejeune, and G. Saporta, editors, Modeles statistiques pour
donnees qualitatives. Editions Technip, Paris.
}
\keyword{datasets}
