% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CorMat}
\alias{CorMat}
\title{Correlation matrix for simulating plsR datasets}
\format{
A data frame with 17 observations on the following 17 variables.
\describe{ \item{y}{a numeric vector} \item{x11}{a numeric
vector} \item{x12}{a numeric vector} \item{x13}{a numeric
vector} \item{x21}{a numeric vector} \item{x22}{a numeric
vector} \item{x31}{a numeric vector} \item{x32}{a numeric
vector} \item{x33}{a numeric vector} \item{x34}{a numeric
vector} \item{x41}{a numeric vector} \item{x42}{a numeric
vector} \item{x51}{a numeric vector} \item{x61}{a numeric
vector} \item{x62}{a numeric vector} \item{x63}{a numeric
vector} \item{x64}{a numeric vector} }
}
\source{
Handmade.
}
\description{
A correlation matrix to simulate datasets
}
\examples{

data(CorMat)
str(CorMat)

}
\references{
Nicolas Meyer, Myriam Maumy-Bertrand et
Frédéric Bertrand (2010). Comparing the linear and the
logistic PLS regression with qualitative predictors: application to
allelotyping data. \emph{Journal de la Societe Francaise de Statistique},
151(2), pages 1-18.
\url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47}
}
\keyword{datasets}
