% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cornell}
\alias{Cornell}
\title{Cornell dataset}
\format{
A data frame with 12 observations on the following 8 variables.
\describe{ \item{X1}{a numeric vector} \item{X2}{a numeric
vector} \item{X3}{a numeric vector} \item{X4}{a numeric
vector} \item{X5}{a numeric vector} \item{X6}{a numeric
vector} \item{X7}{a numeric vector} \item{Y}{response value:
a numeric vector} }
}
\source{
M. Tenenhaus. (1998). \emph{La regression PLS, Theorie et pratique}.
Editions Technip, Paris.
}
\description{
The famous Cornell dataset. A mixture experiment on \code{X1}, \code{X2},
\code{X3}, \code{X4}, \code{X5}, \code{X6} and \code{X7} to analyse octane
degree (\code{Y}) in gazoline.
}
\examples{

data(Cornell)
str(Cornell)

}
\references{
N. Kettaneh-Wold. Analysis of mixture data with partial least
squares. (1992). \emph{Chemometrics and Intelligent Laboratory Systems},
14(1):57-69.
}
\keyword{datasets}
