% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicho.R
\name{dicho}
\alias{dicho}
\title{Dichotomization}
\usage{
dicho(val)
}
\arguments{
\item{val}{A real value}
}
\value{
0 or 1.
}
\description{
This function takes a real value and converts it to 1 if it is positive and
else to 0.
}
\examples{

dimX <- 6
Astar <- 4
(dataAstar4 <- t(replicate(10,simul_data_YX(dimX,Astar))))

dicho(dataAstar4)

rm(list=c("dimX","Astar"))

}
\seealso{
\code{\link{ifelse}}
}
\author{
Frédéric Bertrand\cr
\email{frederic.bertrand@utt.fr}\cr
\url{https://fbertran.github.io/homepage/}
}
\keyword{utilities}
