\name{fowlkes}
\alias{fowlkes}
\docType{data}
\title{Fowlkes dataset}
\description{
A classic dataset from Fowlkes.
}
\usage{data(fowlkes)}
\format{
  A data frame with 9949 observations on the following 13 variables.
  \describe{
    \item{\code{Y}}{binary response}
    \item{\code{MA}}{a numeric vector}
    \item{\code{MW}}{a numeric vector}
    \item{\code{NE}}{a numeric vector}
    \item{\code{NW}}{a numeric vector}
    \item{\code{PA}}{a numeric vector}
    \item{\code{SO}}{a numeric vector}
    \item{\code{SW}}{a numeric vector}
    \item{\code{color}}{a numeric vector}
    \item{\code{age1}}{a numeric vector}
    \item{\code{age2}}{a numeric vector}
    \item{\code{age3}}{a numeric vector}
    \item{\code{sexe}}{a numeric vector}
  }
}
%\source{  ~~ reference to a publication or URL from which the data were obtained ~~}
%\references{  ~~ possibly secondary sources and usages ~~}
\examples{
data(fowlkes)
str(fowlkes)
}
\keyword{datasets}
