% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pineNAX21}
\alias{pineNAX21}
\title{Incomplete dataset from the pine caterpillars example}
\format{
A data frame with 33 observations on the following 11 variables and
one missing value.  \describe{ \item{x1}{altitude (in meters)}
\item{x2}{slope (en degrees)} \item{x3}{number of pines in
the area} \item{x4}{height (in meters) of the tree sampled at the
center of the area} \item{x5}{diameter (in meters) of the tree
sampled at the center of the area} \item{x6}{index of the settlement
density} \item{x7}{orientation of the area (from 1 if southbound to
2 otherwise)} \item{x8}{height (in meters) of the dominant tree}
\item{x9}{number of vegetation strata} \item{x10}{mix
settlement index (from 1 if not mixed to 2 if mixed)}
\item{x11}{logarithmic transform of the average number of nests of
caterpillars per tree} }
}
\source{
Tomassone R., Audrain S., Lesquoy-de Turckeim E., Millier C. (1992).
\dQuote{La régression, nouveaux regards sur une ancienne
méthode statistique}, INRA,
\emph{Actualités Scientifiques et Agronomiques}, Masson,
Paris.
}
\description{
The caterpillar dataset was extracted from a 1973 study on pine
processionary caterpillars. It assesses the influence of some forest
settlement characteristics on the development of caterpillar colonies. There
are k=10 potentially explanatory variables defined on n=33 areas.\cr The
value of x2 for the first observation was removed from the matrix of
predictors on purpose.
}
\details{
These caterpillars got their names from their habit of moving over the
ground in incredibly long head-to-tail processions when leaving their nest
to create a new colony.\cr The \code{pineNAX21} is a dataset with a missing
value for testing purpose.
}
\examples{

data(pineNAX21)
str(pineNAX21)

}
\keyword{datasets}
