% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fowlkes}
\alias{fowlkes}
\title{Fowlkes dataset}
\format{
A data frame with 9949 observations on the following 13 variables.
\describe{ \item{Y}{binary response} \item{MA}{a numeric
vector} \item{MW}{a numeric vector} \item{NE}{a numeric
vector} \item{NW}{a numeric vector} \item{PA}{a numeric
vector} \item{SO}{a numeric vector} \item{SW}{a numeric
vector} \item{color}{a numeric vector} \item{age1}{a numeric
vector} \item{age2}{a numeric vector} \item{age3}{a numeric
vector} \item{sexe}{a numeric vector} }
}
\description{
A classic dataset from Fowlkes.
}
\examples{

data(fowlkes)
str(fowlkes)

}
\keyword{datasets}
