\name{permcoefs.plsRglmnp}
\alias{permcoefs.plsRglmnp}
\title{Coefficients computation for permutation bootstrap}
\description{
  ~~ A (1-5 lines) description of what the function does. ~~
}
\usage{
permcoefs.plsRglmnp(dataRepYtt, ind, nt, modele, family = NULL, 
maxcoefvalues, wwetoile,ifbootfail)
}
\arguments{
  \item{dataRepYtt}{ ~~Explain \code{dataRepYtt} here~~ }
  \item{ind}{ ~~Explain \code{ind} here~~ }
  \item{nt}{ ~~Explain \code{nt} here~~ }
  \item{modele}{ ~~Explain \code{modele} here~~ }
  \item{family}{ ~~Explain \code{family} here~~ }
  \item{maxcoefvalues}{ ~~Explain \code{maxcoefvalues} here~~ }
  \item{wwetoile}{ ~~Explain \code{wwetoile} here~~ }
  \item{ifbootfail}{ ~~Explain \code{ifbootfail} here~~ }
}
\details{
  ~~ More details than the description above ~~
}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}
\references{ ~put references to the literature/web site here ~ }
\author{\enc{Frédéric}{Fr\'ed\'eric} Bertrand\cr
\email{frederic.bertrand@math.unistra.fr}\cr
\url{http://www-irma.u-strasbg.fr/~fbertran/}
}
\note{ ~~some notes~~ 


}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{\donttest{
\dontrun{print("to do")}
}
}\keyword{models}
