% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.plsRglmmodel.R
\name{print.cv.plsRglmmodel}
\alias{print.cv.plsRglmmodel}
\title{Print method for plsRglm models}
\usage{
\method{print}{cv.plsRglmmodel}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{"cv.plsRglmmodel"}}

\item{\dots}{not used}
}
\value{
\code{NULL}
}
\description{
This function provides a print method for the class \code{"cv.plsRglmmodel"}
}
\examples{

data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
print(cv.plsRglm(object=yCornell,dataX=XCornell,nt=10,NK=1,
modele="pls-glm-family",family=gaussian(), verbose=FALSE))
rm(list=c("XCornell","yCornell","bbb"))

}
\references{
Nicolas Meyer, Myriam Maumy-Bertrand et
Frédéric Bertrand (2010). Comparaison de la
régression PLS et de la régression
logistique PLS : application aux données
d'allélotypage. \emph{Journal de la Société Française
de Statistique}, 151(2), pages 1-18.
\url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47}
}
\seealso{
\code{\link{print}}
}
\author{
Frédéric Bertrand\cr
\email{frederic.bertrand@utt.fr}\cr
\url{https://fbertran.github.io/homepage/}
}
\keyword{methods}
\keyword{print}
