% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.plsRmodel.R
\name{summary.cv.plsRmodel}
\alias{summary.cv.plsRmodel}
\title{Summary method for plsR models}
\usage{
\method{summary}{cv.plsRmodel}(object, ...)
}
\arguments{
\item{object}{an object of the class \code{"cv.plsRmodel"}}

\item{\dots}{further arguments to be passed to or from methods.}
}
\value{
An object of class \code{"summary.cv.plsRglmmodel"}.
}
\description{
This function provides a summary method for the class \code{"cv.plsRmodel"}
}
\examples{

data(Cornell)
summary(cv.plsR(Y~.,data=Cornell,nt=10,K=6, verbose=FALSE), verbose=FALSE)

}
\references{
Nicolas Meyer, Myriam Maumy-Bertrand et
Frédéric Bertrand (2010). Comparing the linear and the
logistic PLS regression with qualitative predictors: application to
allelotyping data. \emph{Journal de la Societe Francaise de Statistique},
151(2), pages 1-18.
\url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47}
}
\seealso{
\code{\link{summary}}
}
\author{
Frédéric Bertrand\cr
\email{frederic.bertrand@utt.fr}\cr
\url{https://fbertran.github.io/homepage/}
}
\keyword{methods}
\keyword{print}
